refactor(all)!: internal local package with general cfg
