+++ setup.cfg
@@ -27,4 +27,8 @@ skip =
     migrations/,
     */migrations/*,
     */tests/*
-src_paths = api,consumer,ws
+src_paths = 
+    internal,
+    api,
+    consumer,
+    ws,
